Extracting TagList to a separate composable.
