+++ app/src/main/java/in/developingdeveloper/timeline/taglist/ui/components/TagList.kt
@@ -1,16 +1,25 @@
 package `in`.developingdeveloper.timeline.taglist.ui.components
 
 import android.content.res.Configuration
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.LinearProgressIndicator
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import `in`.developingdeveloper.timeline.core.ui.theme.TimelineTheme
+import `in`.developingdeveloper.timeline.taglist.ui.models.TagListViewState
 import `in`.developingdeveloper.timeline.taglist.ui.models.UITag
 
 @Composable
@@ -27,6 +36,46 @@ fun TagList(
             TagListItem(tag = tag)
         }
     }
+}
+
+@Composable
+fun TagList(viewState: TagListViewState) {
+    when {
+        viewState.isRefreshing -> {
+            LinearProgressIndicator(modifier = Modifier.fillMaxWidth())
+        }
+
+        viewState.isLoading -> {
+            CircularProgressIndicator(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .wrapContentSize(),
+            )
+        }
+
+        else -> {
+            Box(modifier = Modifier.height(4.dp))
+        }
+    }
+
+    viewState.message?.let { message ->
+        CenterText(message)
+    }
+
+    if (viewState.shouldDisplayTags) {
+        TagList(tags = viewState.tags)
+    }
+}
+
+@Composable
+private fun CenterText(text: String) {
+    Text(
+        text = text,
+        style = MaterialTheme.typography.titleMedium,
+        modifier = Modifier
+            .fillMaxSize()
+            .wrapContentSize(),
+    )
 }
 
 @Preview(
